requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
Pillow==9.4.0
python-dotenv==1.0.0
pytz==2023.2
sqlparse==0.4.3
tzdata==2023.2
requests
